+++ client/src/client/mod.rs
@@ -11,7 +11,6 @@ use std::{
     time::Duration,
 };
 
-use iota_pow::providers::{NonceProvider, NonceProviderBuilder};
 use iota_types::block::{output::RentStructure, protocol::ProtocolParameters};
 #[cfg(not(target_family = "wasm"))]
 use tokio::runtime::Runtime;
@@ -101,25 +100,6 @@ impl Client {
     /// Create the builder to instantiate the IOTA Client.
     pub fn builder() -> ClientBuilder {
         ClientBuilder::new()
-    }
-
-    /// Gets the miner to use based on the Pow setting
-    pub fn get_pow_provider(&self) -> impl NonceProvider {
-        let local_pow: bool = self.get_local_pow();
-        #[cfg(target_family = "wasm")]
-        let miner = crate::api::wasm_miner::SingleThreadedMiner::builder()
-            .local_pow(local_pow)
-            .finish();
-        #[cfg(not(target_family = "wasm"))]
-        let miner = {
-            let mut miner = crate::api::miner::ClientMiner::builder().with_local_pow(local_pow);
-            if let Some(worker_count) = self.pow_worker_count {
-                miner = miner.with_worker_count(worker_count)
-            }
-            miner.finish()
-        };
-
-        miner
     }
 
     /// Gets the network related information such as network_id and min_pow_score