+++ types/tests/block.rs
@@ -1,13 +1,7 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use iota_pow::{
-    providers::{
-        miner::{Miner, MinerBuilder},
-        NonceProviderBuilder,
-    },
-    score::PoWScorer,
-};
+use iota_pow::{miner::get_miner, score::PowScorer};
 use iota_types::block::{
     parent::Parents,
     payload::{Payload, TaggedDataPayload},
@@ -23,28 +17,21 @@ use iota_types::block::{
 use packable::{error::UnpackError, PackableExt};
 
 #[test]
-fn pow_default_provider() {
-    let min_pow_score = protocol_parameters().min_pow_score();
-    let block = BlockBuilder::<Miner>::new(rand_parents())
-        .finish(min_pow_score)
-        .unwrap();
+fn default_finish_zero_nonce() {
+    let block = BlockBuilder::new(rand_parents()).finish().unwrap();
 
-    let block_bytes = block.pack_to_vec();
-    let score = PoWScorer::new().score(&block_bytes);
-
-    assert!(score >= min_pow_score as f64);
+    assert!(block.nonce() == 0);
 }
 
 #[test]
 fn pow_provider() {
     let min_pow_score = protocol_parameters().min_pow_score();
     let block = BlockBuilder::new(rand_parents())
-        .with_nonce_provider(MinerBuilder::new().with_num_workers(num_cpus::get()).finish())
-        .finish(min_pow_score)
+        .finish_nonce(get_miner(min_pow_score))
         .unwrap();
 
     let block_bytes = block.pack_to_vec();
-    let score = PoWScorer::new().score(&block_bytes);
+    let score = PowScorer::new().score(&block_bytes);
 
     assert!(score >= min_pow_score as f64);
 }
@@ -52,13 +39,13 @@ fn pow_provider() {
 #[test]
 fn invalid_length() {
     let res = BlockBuilder::new(Parents::new(rand_block_ids(2)).unwrap())
-        .with_nonce_provider(42)
+        .with_nonce(42)
         .with_payload(
             TaggedDataPayload::new(vec![42], vec![0u8; Block::LENGTH_MAX - Block::LENGTH_MIN - 9])
                 .unwrap()
                 .into(),
         )
-        .finish(protocol_parameters().min_pow_score());
+        .finish();
 
     assert!(matches!(res, Err(Error::InvalidBlockLength(len)) if len == Block::LENGTH_MAX + 33));
 }
@@ -66,9 +53,9 @@ fn invalid_length() {
 #[test]
 fn invalid_payload_kind() {
     let protocol_parameters = protocol_parameters();
-    let res = BlockBuilder::<Miner>::new(rand_parents())
+    let res = BlockBuilder::new(rand_parents())
         .with_payload(rand_treasury_transaction_payload(protocol_parameters.token_supply()).into())
-        .finish(protocol_parameters.min_pow_score());
+        .finish();
 
     assert!(matches!(res, Err(Error::InvalidPayloadKind(4))))
 }
@@ -111,9 +98,7 @@ fn unpack_invalid_remaining_bytes() {
 #[test]
 fn pack_unpack_valid() {
     let protocol_parameters = protocol_parameters();
-    let block = BlockBuilder::<Miner>::new(rand_parents())
-        .finish(protocol_parameters.min_pow_score())
-        .unwrap();
+    let block = BlockBuilder::new(rand_parents()).finish().unwrap();
     let packed_block = block.pack_to_vec();
 
     assert_eq!(packed_block.len(), block.packed_len());
@@ -132,8 +117,8 @@ fn getters() {
 
     let block = BlockBuilder::new(parents.clone())
         .with_payload(payload.clone())
-        .with_nonce_provider(nonce)
-        .finish(protocol_parameters.min_pow_score())
+        .with_nonce(nonce)
+        .finish()
         .unwrap();
 
     assert_eq!(block.protocol_version(), protocol_parameters.protocol_version());